Make some changes to server file structure and write server.js
