+++ server/routes.js
@@ -1,0 +1,3 @@
+const express = require('express');
+
+module.exports = {}
